moved API views and related functions related to Service page to a separate file
